Continuing to fix up tests
