+++ test/redis/test_redis.cpp
@@ -2,7 +2,6 @@
 
 #include "utils.h"
 
-#include <scheduler/scheduler.h>
 #include <util/util.h>
 #include <algorithm>
 
@@ -166,86 +165,6 @@ namespace tests {
         std::string nonExistKey = "blahblah";
         std::vector<uint8_t> actual = redisQueue.get(nonExistKey);
         REQUIRE(actual.empty());
-    }
-
-    void checkCallRoundTrip(bool isAsync, bool isSuccess) {
-        cleanSystem();
-
-        // Request function
-        message::Message call;
-        std::string funcName = "my func";
-        std::string userName = "some user";
-        std::string inputData = "blahblah";
-
-        call.set_function(funcName);
-        call.set_user(userName);
-        call.set_inputdata(inputData);
-        call.set_isasync(isAsync);
-        call.set_success(isSuccess);
-
-        // Check resultkey not set initially
-        REQUIRE(!call.has_resultkey());
-
-        scheduler::Scheduler &sch = scheduler::getScheduler();
-        std::string queueName = sch.callFunction(call);
-
-        // Get the next call
-        scheduler::MessageQueue messageQueue;
-        message::Message actual = messageQueue.nextMessage(queueName);
-
-        REQUIRE(funcName == actual.function());
-        REQUIRE(userName == actual.user());
-        REQUIRE(inputData == actual.inputdata());
-        REQUIRE(isAsync == actual.isasync());
-        REQUIRE(isSuccess == actual.success());
-
-        // Check result key has now been set
-        REQUIRE(actual.has_resultkey());
-    }
-
-    TEST_CASE("Test redis round trip sync fail", "[redis]") {
-        checkCallRoundTrip(false, false);
-    }
-
-    TEST_CASE("Test redis round trip sync success", "[redis]") {
-        checkCallRoundTrip(false, true);
-    }
-
-    TEST_CASE("Test redis round trip async fail", "[redis]") {
-        checkCallRoundTrip(true, false);
-    }
-
-    TEST_CASE("Test redis round trip async success", "[redis]") {
-        checkCallRoundTrip(true, true);
-    }
-
-    TEST_CASE("Test redis reading and writing function results", "[redis]") {
-        Redis &redisQueue = Redis::getQueue();
-        redisQueue.flushAll();
-
-        // Write some function result
-        message::Message call;
-        call.set_function("my func");
-        call.set_user("some user");
-        call.set_resultkey("function 123");
-
-        scheduler::MessageQueue messageQueue;
-        messageQueue.setFunctionResult(call, true);
-
-        // Check result has been written to the right key
-        REQUIRE(redisQueue.listLength("function 123") == 1);
-
-        // Check that some expiry has been set
-        long ttl = redisQueue.getTtl(call.resultkey());
-        REQUIRE(ttl > 10);
-
-        // Check retrieval method gets the same call out again
-        message::Message actualCall2 = messageQueue.getFunctionResult(call);
-
-        REQUIRE("my func" == actualCall2.function());
-        REQUIRE("some user" == actualCall2.user());
-        REQUIRE("function 123" == actualCall2.resultkey());
-        REQUIRE(actualCall2.success());
     }
 
     TEST_CASE("Test setnxex", "[redis]") {
@@ -466,7 +385,7 @@ namespace tests {
         std::string setA = "set_a";
         std::string setB = "set_b";
 
-        const std::vector<std::string> actualA = redisQueue.sdiff(setA, setB);
+        const std::set<std::string> actualA = redisQueue.sdiff(setA, setB);
         REQUIRE(actualA.empty());
 
         std::string valueA = "val_a";
@@ -481,11 +400,8 @@ namespace tests {
         redisQueue.sadd(setB, valueC);
         redisQueue.sadd(setB, valueD);
 
-        std::vector<std::string> actualB = redisQueue.sdiff(setA, setB);
-        std::vector<std::string> expected = {valueA, valueB};
-
-        std::sort(actualB.begin(), actualB.end());
-        std::sort(expected.begin(), expected.end());
+        std::set<std::string> actualB = redisQueue.sdiff(setA, setB);
+        std::set<std::string> expected = {valueA, valueB};
 
         REQUIRE(actualB == expected);
     }
@@ -497,7 +413,7 @@ namespace tests {
         std::string setA = "set_a";
         std::string setB = "set_b";
 
-        const std::vector<std::string> actualA = redisQueue.sinter(setA, setB);
+        std::set<std::string> actualA = redisQueue.sinter(setA, setB);
         REQUIRE(actualA.empty());
 
         std::string valueA = "val_a";
@@ -515,11 +431,8 @@ namespace tests {
         redisQueue.sadd(setB, valueD);
         redisQueue.sadd(setB, valueE);
 
-        std::vector<std::string> actualB = redisQueue.sinter(setA, setB);
-        std::vector<std::string> expected = {valueA, valueC};
-
-        std::sort(actualB.begin(), actualB.end());
-        std::sort(expected.begin(), expected.end());
+        std::set<std::string> actualB = redisQueue.sinter(setA, setB);
+        std::set<std::string> expected = {valueA, valueC};
 
         REQUIRE(actualB == expected);
     }
